Displaying refs for fetch tasks
Fetch tasks string representation is currently showing a unique id
alongside the project the fetch is for.

Whilst this is useful it does not tell which refs are actually being
fetched.

Add the refs to the string representation of the fetch tasks allows to
have a better understanding of what is going on when looking at a
show-queue or when logging the tasks themselves.

Change-Id: I53f9e5da54d01ccbda772cf2c49b3b87b3628ca7