Log fetch tasks when graceful shutdown fails
When shutting down the pull-replication, an orderly shutdown is
initiated and the plugin waits for the fetch tasks to be completely
drained, for a period of time.

If an graceful shutdown fails however, we know nothing about which fetch
tasks were interrupted.

It would be useful to know which tasks fail, in which status they were
and which refs they were fetching.

Log pending, in-flight and never executed fetch tasks in the error_log
file, to help troubleshooting as well as facilitating admins to
understand which refs might need manual fetching.

Change-Id: Iafe33158959005f1a2e4a44e95f818a7e417ca53