Allow subscribing to the topic with custom group id
In the past, when utilizing pull-replication as a standalone solution
in conjunction with a broker (inclusive of plugins such as events-kafka,
events-aws-kinesis, events-gcloud-pubsub), consumer configurations
relied on the groupId specified within the events-* plugin configuration.
In cases where no groupId was explicitly defined, the system defaulted to
using the value "gerrit.instance-id."

Thanks to a new feature in events-broker, since v3.4.8-4-gd9f859f,
it is possible to define a different group id by configuring
"replication.eventBrokerGroupId" in the replication.conf.

NOTE: This feature is applicable only if the the events-* plugin
implements the ExtendedBrokerApi interface, which supports the ability
to subscribe to a topic with a custom group id.

Bug: Issue 299327285
Change-Id: I33d056ebd3e80b6a4c9d8eb02a0c2c1b591e8eb5