Allow subscribing to the topic with custom group id
In the past, when utilizing pull-replication as a standalone solution
in conjunction with a broker (inclusive of plugins such as events-kafka,
events-aws-kinesis, events-gcloud-pubsub), the consumer's groupId was
defined either within the events-* plugin configuration or, in the
absence of configuration, as "gerrit.instance-id."

Now, thanks to this new feature, the pull-replication plugin offers
the option to define a group id when the property
"replication.eventBrokerGroupId" is specified in the replication.conf
file. Additionally, this feature is applicable when the events-* plugin
implements the ExtendedBrokerApi interface.

Bug: Issue 299327285
Change-Id: I33d056ebd3e80b6a4c9d8eb02a0c2c1b591e8eb5