Allow subscribing to the topic with custom group id
Previously, when the pull-replication plugin was set up with events-*
plugins (such as events-kafka, events-aws-kinesis,events-gcloud-pubsub),
consumer configurations depended on the groupId specified in the
events-* plugin configuration. If no groupId was provided, it defaulted
to using the value "gerrit.instance-id."

Now, thanks to this new feature, the pull-replication plugin offers
the option to define a group id when the property
"replication.eventBrokerGroupId" is specified in the replication.conf
file. Additionally, this feature is applicable when the events-* plugin
implements the ExtendedBrokerApi interface.

Bug: Issue 299327285
Change-Id: I33d056ebd3e80b6a4c9d8eb02a0c2c1b591e8eb5