Stop tracking the replication end2end latency metric on the sender
The `replication_end_2_end_latency` metric is being populated by both
the sender of the fetch request and the recipient, which makes it
confusing for someone who is looking the metric to understand what
value is appropriate.

Stop tracking this metric on the sender, but calculate the time it
took and log it for diagnostic purposes.

Issue: 40015571
Change-Id: I665a23c288f293b947366f5b0d172af604bac8dc