Fix pull-replication plugin startup which was always failing
The startup of the pull-replication plugin was constantly failing
because the DynamicItem brokerApi was never an instance of
ExtendedBrokerApi, which is expected: they are always different
types and represent different classes.

The error was not spotted during review because of the incorrect
naming of the DynamicItem as eventsBroker, which is a bad name
because it isn't the broker but just a DynamicItem pointing to
a broker.

Rename eventsBroker as eventsBrokerDi and check for the real
brokerApi being an instance of ExtendedBrokerApi.

Change-Id: I42f1738dcffcf438e5bd2a0b6978e098f838fdad