Run synchronous fetch as a single FetchOp for all refs in a batch
Following the introduction of the batch-fetch api [1], the work on the
synchronous fetch side was outstanding. The payload of the batch fetch
contains a list of all the refs in the batch, for which we want to
execute a single git fetch operation. With regards to the async operations
all the scheduling would end up in the same replication task as they have
in common the same URI and repository.

[1] https://gerrit-review.googlesource.com/c/plugins/pull-replication/+/378496

Bug: Issue 303112557
Change-Id: Ifd2a28a6337f762d89d744934f44b7f2b47e6e6a