Remove white-box unit tests on synchronous FetchCommand
The purpose of unit testing should be verifying that a
component respects its contract with its user through an
API. Some of the tests of the synchronous fetch replication
did instead verify the internal implementation detail of
the replication happening through a zero-delayed execution
of a scheduled task.

Remove all white-box tests which would be misleading in
giving a false perception of confidence that the component
was working.

Change-Id: I5c0b3ee4ceb0456a0b5d12463054afcfe8a4ee0f