Invoke the service method when using ProjectInitializationAction
The Java class ProjectInitializationAction, which extends HttpServlet,
now utilizes the HttpServlet API's service method within the context
of the PullReplicationFilter, replacing the previous usage of the
initProject method.

The objective of this modification is to encapsulate the
'init project' use case within the prescribed pattern:
Request -> PullReplicationFilter -> ProjectInitializationAction.
In this pattern, the filter's sole responsibility is to intercept,
process the request, and forward it to ProjectInitializationAction,
where the core logic is implemented in accordance with the HttpServlet
engine conventions.

Change-Id: I6e3c25216489d71241e1563eee219976fdfaca1c