Extend FetchApiClient interface to init project with config
This change extends the FetchApiClient interface by adding the
ability to replicate the "init project" operation along with
its configuration.

The new feature involves sending an HTTP PUT request to the
endpoint:  pull-replication/init-project-with-config/<project_name>.git,
including the configuration (refs/meta/config) as part of the
request body.

It's important to note that this modification has no impact on
the critical path, and therefore, there is no risk of disrupting
any functionality.

Bug: Issue 296854545
Change-Id: I01fc49c5327b48efc7f3292c313a067076b4b0ed