Add JGit client test for unset mirror
The unset mirror test for the CGit client was added in a previous
change, but change 389376, which added this feature to JGit, didn't
include this.

Add test for the scenario where the mirror flag is unset, i.e. don't
remove remotely tracked refs. This provides the same coverage for this
feature for CGit and JGit.

Bug: Issue 40015295
Change-Id: I5b329b73c291536abc6d44d0d76c3f73c4489dbc