Add endpoint that allows init project with configuration
This change enhances the capabilities of the Java class
PullReplicationFilter, enabling it to intercept requests directed
to the "pull-replication/init-project-config/<project_name>.git"
endpoint and forward them to the ProjectInitializationWithConfigAction
(an HttpServlet). The responsibilities of the
ProjectInitializationWithConfigAction are as follows:

- Creating the project within the file system.
- Parsing the payload into a RevisionsInput object.
- Validating the RevisionsInput object.
- Executing the apply object operation on the RevisionsInput object.
- Caching the project.

To preserve backward compatibility with Gerrit instances running older
versions, the previous endpoint for enabling "init project" remains
unchanged.

Bug: Issue 296854545
Change-Id: Icbff1a19a4a1a2b498d16300928efc862b40739e