Revise the "init project" endpoint to manage project configuration
This change expands the functionalities of the Java class
"ProjectInitializationAction" to process the payload when it is
present. When the payload is available, the logic will be as follows:

- Create the project within the file system.
- Parse the payload into a RevisionsInput object.
- Validate the RevisionsInput object.
- Execute the apply objects operation on the RevisionsInput object.
- Cache the project.

It's essential to emphasize that this modification will have no impact
on earlier Gerrit instances that use older versions of the
pull-replication plugin when replicating "init project" without
including the project configuration in the payload.

Bug: Issue 296854545
Change-Id: Ib737003754f68db7484f3b0fc7a7c9e12e4073f5