Reverse the order of storing commit, tree and blobs
Apply-object inserts to the repository commits, trees and blobs.
These writes are no transactional this means that if any of these
writes fails other are not reverted. This can cause repository
corruption because we end up with a commit pointing to non existing
tree or blob.
To make sure that repository is not corrupted make sure that blobs
are inserted before trees and trees are inserted before commits.
With this order even if the insertion fails it will keep repository
healthy and next git GC will clean up dangling objects.

Bug: Issue 306464772
Change-Id: I761e84d5e518e3c3b8ab9c151996149e5d466378