Remove events from beforeStartupEventsQueue in-memory queue
Before pull-replication plugin is fully loaded it can receive
events. These events are stored in an in-memory queue called
`beforeStartupEventsQueue`. After plugin is fully loaded those
events are processed but never removed from the queue. This can
cause issue processing the same event multiple times.

Replication done with apply-object may not be idempotent if
the event generated at a point in time is processed later.

To avoid that remove event from the queue and then process it.

Bug: Issue 307592553
Change-Id: I946deebe15b92bad00e1127e738ef2540ee78983