Set `application/json; charset=utf-8` as content-type when fetch
This modification involves adjusting the content type for the fetch
operation. It now specifies the media type as "application/json" and
includes the character encoding as "charset=utf-8." Previously, the
character encoding was not defined, which is less than ideal because
it's essential for ensuring the accurate interpretation of textual data.

This change is non-disruptive as the server-side currently does not
validate the content type. The validation of content types will be
addressed in a subsequent update.

Change-Id: I66c3ca93527a944f5f49f15b22d90addfa58b628