Force async fetch when replying beforeStartupEventsQueue events
Replying beforeStartupEventsQueue events can use apply-object
operation. This can cause the revert of data and a split-brain.
Apply-object actions are designed to be synchronous and near-real-time.
When they are triggered asynchronously, like in this case, they will
bring the branch back in time, because they use the SHA1 of the event
instead of the one on the local ref on the filesystem.

To avoid those issues skip apply-object call.

It is tough to write a test for this specific condition and verifying
that the fetch happened instead of apply-object.

Bug: Issue 307574381
Change-Id: I9ce8cf05d1d4edd7e8e0e76b8f4762e37e5b10db