Move cache eviction before the applyObjects when init a new project
When a new project is created, the cache could have a negative hit
and the initialisation of a new project with the associated
refs/meta/config may fail because of the cached failure.

Evicting after the applyObjects() call is not useful in case of negative
caching because it would risk also to fail.

Change-Id: I7c77e0b3efa52c35248e3417d01348aff6a879e1