Fix tests broken with the upgrade of Mockito framework
With the change 390986, Mockito framework was upgraded from 3.3.3
to 5.6.0 and because of that some tests that stop working. This
change reqires minor refactoring to fix those tests, as outlined
in the following points:

- The method `verifyZeroInteractions` is no longer supported in the
class `org.mockito.Mockito` [1]. Instead, it can be substituted with
`verifyNoInteractions`.

- In Mockito version 3.3.3, the execution of real code occurred in
methods without defined expectations, and some tests were designed
accordingly. However, with Mockito 5.6.0, when methods are invoked
without specified expectations, the framework will now appropriately
return either null, a primitive/primitive wrapper value, or an empty
collection. As a result, minor refactoring may be necessary.

- The handling of Mockito expectations in methods declared with varargs
has been enhanced to encompass all the arguments.

References:

[1] https://github.com/mockito/mockito/blob/v5.6.0/src/main/java/org/mockito/Mockito.java#L2699

Change-Id: Ic3d8641c4ca6b485212bdaa55d359a4cc4bb4741