Don't require remote.<name>.fetch on primary side
The pull-replication running in primary mode doesn't perform fetch from
a replica, and, therefore it doesn't need `url` and `fetch` properites
from the `remote` configuration`.

The remote.<name>.url is already ignored in the primary node, thanks to
change 373627[1]. This change ensures that `ConfigInvalidExecption` will
not be thrown on the primary node when remote.<name>.fetch option is
missing.

[1]
https://gerrit-review.googlesource.com/c/plugins/pull-replication/+/373627

Bug: Issue 40015281
Change-Id: I650f9c796534e0af908d6adc6e02e0af92cbd866