Remove events from beforeStartupEventsQueue after fire()
The plugin was removing the event before firing it from the
beforeStartupEventsQueue. This could have caused a message
loss in case the configuration was changed before firing the event.

Now after the plugin is fully loaded each event is
removed from the the in-memory queue after being fired.

Follow up of change 391054.

Change-Id: I4ccef01f42dc11c5df84ca8dfa231e6e895d8a2d