Index change asynchronously upon refs replicated event
Do not block the replication thread for performing a change
indexing, as it may slowdown the incoming operation on the
sender side.

When the refs are replicated through apply-object or sync
replication, the incoming thread is blocked until all listeners
are complete. The reindexing of a change may take a considerable
amount of time, especially when there are Prolog rules involved.

The ref is replicated therefore there is no need to block the
incoming operation waiting for the reindex to complete.

Change-Id: I49279262ea68bb6118ea76ff3123a7538dfe72fc