Accept remotes without `fetch` option on primary
The primary replication node only uses `apiUrl` to communicate with
secondary nodes. This means that remote confiugrations without the `url`
and `fetch` options should be valid.

The `remote.<name>.url` is already considered not-required option on the
primary.

With this change, the remote configurations on the primary node without
`fetch` option will also be considered valid.

Bug: Issue 40015281
Change-Id: I701dd82481b28b9a051f97b2a461ebf9b96b0025