Notify listeners when a new project is created via replication
Assume multi-site setup with nodes 1 and 2 in site A and node 3 in site
B. Nodes 1 and 2 run the high-availability plugin and they share the
filesystem (eg over NFS). Also, node 1 is the remote target in node 3's
replication.config. When a project is created in node 3, it is possible
that the "project-created" stream event will be consumed by node 2.
This means:
- The project will be created in node 1 via pull-replication.
- Node 2 will attempt to create the project as a result of consuming the
stream event. The project creation will fail, as the project already
exists on the filesystem.
- If the project-list-update event is consumed by node 1, then the
project will not be added to the project_list cache in node 2, and
therefore it won't be visible in the UI for a little while.

To fix this, when a project is created via replication, notify all
`NewProjectCreatedListener`s. High-availability defines a handler which
listens to such events, and forwards a `ProjectListUpdateCache` event
to the other master.

Bug: Issue 40015518
Change-Id: I1b22bf1fb0c7654d4ad2493eb0a5a0e568500e53