Notify listeners when a new project is created via replication
In a complex multi-site with high-availability setup, the plugins that
are reacting on the creation of a project were not able to detect the
ones created via replication, which lead to unpredictable consequences.

To fix this, when a project is created via replication, notify all
`NewProjectCreatedListener`s of the operation that took place in the
local node.

Bug: Issue 40015518
Change-Id: I1b22bf1fb0c7654d4ad2493eb0a5a0e568500e53