Fix ApplyObjectActionIT flakiness due to the wrong ref used
The testing of the apply-object was using a patch-set which
was missing its parent object, because it was done on a separate
repository created from scratch with a different master.

Use consistently refs/heads/master on its first commit when
testing apply-object and make sure that the destination repo
does not have it so that the apply-object would not have any
reason to fail.

Change-Id: I0dba27505a83b62300a84903f4a520c205f09eeb