Pull-replication plugin should warn about inconsistent timeouts
The pull-replication uses the `connectionTimeout` for all the
socket-related timeouts, including the read/write timeouts. At the
moment, the only way to assure a reliable replication is to have a
`connectionTimeout` > `timeout`.

The replication plugin should provide a warning when an
inconsistent configuration is found in any of the remotes.

Bug: Issue 40014561
Change-Id: I04f18d5307de2333b8a50294c557bdd0213410a8