Set remote.NAME.fetch default value on replicas
Mimic the behaviour of replicaton plugin which uses `refs/*:refs/*` as
default value for `remote.NAME.fetch`. This only apply on replicas where
the `remote.NAME.fetch` is actually used. On the primary node this
configuration option will be ignored, meaning that we don't need a
default value for it.

Change-Id: Iec2b0402103f1dee861626eacb79838f0cdb07a7