Use native BatchInput for the sync batch-fetch REST-API
Use the BatchInput data type for the batch-fetch REST-API
for reducing the payload size, keeping the correct granularity
of the batch ref-update data and disallowing at protocol
level the possibility to split a batch into sync and async
executions.

Also avoid scheduing a batch as a series of individual
fetch operations of the individual refs, as a follow-up
of the review of Ifd2a28a6.

Change-Id: I1b8fcb0b803cd441f4fac21940c205767d0eb3a9