Read heathcheck settings from plugin configuration
Some settings are specific to the external healthcheck; the list of
repos and the periodOfTime. These settings are defined in the plugin's
configuration file.

Read custom settings from the plugin's config file. "Core" settings
like the `enabled` flag and the timeout are still defined in the
healthcheck plugin's config file.

Bug: Issue 312895374
Change-Id: I8454b19770e4cba986b1e33bddf15f00bc7a8f70