Fix: pull-replication could not be installed without replication
Change 396999 [1] introduced support for overriding replication
configuration through the `ReplicationConfigOverrides` interface, for
which a dynamic item implementation can be provided.

This, in turn, broke the pull-replication which needed to be adapted to
support the new `ConfigResource` and `ConfigResourceOverrides` objects
(this was addressed in change 400819 [2]).

Whilst the adaptation fixed the tests by explicitly installing the push
replication `ApiModule()` in `TestModule()`s, it overlooked adapting the
`PullReplicationModule` itself assuming that the
`ConfigResourceOverrides` DyamicItem was always provided by the
replication plugin itself.

This was true, as long as the `replication` plugin was installed
alongside the `pull-replication`, however in cases where the
`replication` plugin was not installed, the `pull-replication` would
fail to load due to the lack of `ConfigResourceOverrides` DynamicItem
binding.

Fix this by instructing the `PullReplicationModule` to do an explicit
binding of the `ConfigResourceOverrides` where not already present.

[1] https://gerrit-review.googlesource.com/c/plugins/replication/+/396999
[2] https://gerrit-review.googlesource.com/c/plugins/pull-replication/+/400819

Bug: Issue 321784346
Change-Id: Ie06f2418730cbf6a595ebb3e99c6b7669a729899