Itntroduce the `remeote.NAME.fetchAllEvery` configuration param
Currently, there are two methods for obtaining repository updates in
pull-replication:
* Apply objects - This transfers updated refs and objects via the REST
  API, eliminating the need for git fetch.
* Trigger git fetch - This is used when objects exceed the REST API
  payload limit, necessitating a git fetch.

The parameter (if configured and when implemented) enables the third,
time-based (every `n` seconds) method, that is independent of the
REST API. It will periodically invoke git fetch to detect and retrieve
new data as it becomes available. Note that it is meant for the remote
that doesn't offer any events or webhooks to provide information about
new data.

The default parameter value is `0s` which means that periodical fetch
is disabled. Note that larger time units (`m`, `h`, etc...) can be
used to configure it conveniently.

Bug: Issue 322146240
Change-Id: I23fc5e23aabff2dd0f053de235245614cf4706c9