Remove already bound classes from PullReplicationModule
The change I8003ec6c, for exposing API to update replication remotes',
has also rearranged the guice bindings, moving some of them from the
main `ReplicationModule` to `ReplicationConfigModule`. This then lead to
double bindings in the `PullReplicationModule` as reported in bug
330495476.

As the `PullReplicationModule` is already installing the
`ReplicationConfigModule`, we can safely remove bidnings provided by the
config module.

Bug: Issue 330495476
Change-Id: I22bbdeb4901f703881e3e8617aa1400873151882