Fix compilation and test errors after remotes' API merge
Recently the `replication` plugin got new public API to modify the
remotes' configuraiton. It also required some reshuffling in the
`ReplicationConfigModule` and `ReplicationModule`. Aswell as addition of
a new method to `ReplicationConfigOverrides`.

The above caused compilation and binding errors. To fix them, bindings
moved to the `ReplicationConfigModule` were removed plus, we use now a
module overriding to bind our custom `ConfigParser` implementation.

THis also adds a no-op implementation of
`ReplicationConfigOverrides.update(Config)` method in tests.

Bug: Issue 330495476
Change-Id: I22bbdeb4901f703881e3e8617aa1400873151882