Do not memoize the FetchOne computation of refSpecs.
The processing of the delta-refs into refSpecs cannot
be memoized into the FetchOne instance because it would
risk to execute the fetch operation on an stale delta,
causing the loss of precious replication refs.

FetchOne can be rescheduled because of locking issues
or concurrency with other threads. During the rescheduled
retry there could be more refs added to the delta that
would need to be also taken into account.

Bug: Issue 337338030
Change-Id: I0a041a1d5896c619ec4184aea61fb3035f70de02