FetchOne: refresh deltaRefSpecs every time that delta changes
Replace the memoizing of the computed deltaRefSpecs with a more
dynamic recalculation free of locks that would reuse an AtomicReference
to an in-memory list across calls, whenever the delta-refs has not
changes.

If the delta is mutated, for example when a replication task is
rescheduled and receives more refs or its associated global-refdb
changes, recalculate the deltaRefs and take into consideration
the new mutated delta.

Bug: Issue 337338030
Change-Id: I24d70b796b870d90fc487cd31445d8249c5801bd