FetchOp: block incoming delta refs when fetch is in progress
Adding extra refs to fetch whilst the fetch has started already
isn't useful and would only cause the refs to be added to be
lost and missing its replication action.

Also, once the FetchOp has completed, do not allow the Soource
to add extra refs to it, even if the FetchOp was taken from
the pending queue.

Protect with synchronised for concurrency between the addition
of refs and the completion state.

Change-Id: I5f1cd9824be19019fd6843c6130d6f0056f18191