Stop memoizing fetch refSpecs
Change I29507376e introduced the memoization fetch refsSpecs.

This was intended to avid expensive, repeated delta computations for
fetches without an explicit delta to fetch (for example via the
`pull-replication start` SSH command).

The problem with this approach however is that it doesn't account for
the fact that delta is a mutable entity: `FetchOne` could be rescheduled
for whatever reason, consolidating new refs into a new delta.

If now `FetchOne` tries to replicate again based on the new delta refs,
the memoizing causes the return of the previous delta value, causing the
effective loss of the refs to replicate.

Rather than memoizing the delta itself, we now just memoize the
expensive computation of comparing local refs vs remote refs,
exclusively used during when fetching from source without any explicit
refs (i.e. no delta).

Bug: 337338030
Change-Id: I06442a66896205258b2b0411f74855465c333864