Ban synchronous ref-delete via apply-object(s)
Ref deletion is a non-fast-forward operation and it may
have different outcome with other forced operations when
not executed in the correct order.

Ban the use of ref-delete via apply-object returning a 400
and force the client to run an asynchronous fetch instead
which would also go through global-refdb validation.

Example: a forced update and a delete cannto be swapped

Bug: Issue 367651566
Change-Id: I94d2850c4aad7573fabfd83bd361e1e683b8cb19