Ban synchronous ref-delete via apply-object(s)
Ref deletion is a non-fast-forward operation and it may
have different outcome with other forced operations when
not executed in the correct order.

Ban the use of ref-delete via apply-object returning a 400
and force the client to run an asynchronous fetch instead
which would also go through global-refdb validation.
For example, a forced update and a delete cannto be swapped.

Bug: Issue 367651566
Change-Id: I978c3079fbb9db4ea5b4fcbb63df711430ad0ce8