Introduce FetchRefSpec over the whole replication queue processing
Do not rely on simple String ref renames for tracking
replication tasks and keep the full RefSpec all throughout
the processing, expanding it further using the configured
RefSpec. This allows to represent further fetch operations like
refs deletion, which is leveraged as a follow-up of this
change.

Introduce FetchRefSpec as a subclass of RefSpec with additional
utility methods for translating from/to a refname and a plain
RefSpec.

Change-Id: I9141eade50058c128c0f14e33fec9d755fea16c7