Delete refs as part of the replication queue execution
Do not delete refs synchronously in any case and use
instea the same queue for the regular fetch operations.
This allows to keep the ordering of the delete refs
against any other fetch.

Also remove the now unused async DeleteRefJob because the
deletion is managed in the same replication queue as any
other ref.

Bug: Issue 367651566
Change-Id: I3b43c0c5fb5e9e1189bf0e3990b7981e3e68cd92