Add tests for asynchronous fetch replication
Make sure that ref/batch-ref updates are tested
in asynchronous fetch replication as well, managing also
the situation where replication tasks are retrying and
update/delete need to be applied in the correct order.
Test timeout is set to a maximum of 2-times the replication
delay, rather than 2k-times which was leading to
tests starvation after over 2h (4000 sec) of waiting for a
condition to be completed.

Also fix an issue where async fetch were not executed
in background but the incoming REST-API was blocking until
the fetch completion, which was unintentional for async
operations.

Change-Id: I0984ac1af805ac79b3c4dd5e96cf122f671ba3b6